#201: put image fetch endpoint under context-path 'raw'
