+++ src/main/scala/no/ndla/imageapi/ImageApiProperties.scala
@@ -25,7 +25,9 @@ object ImageApiProperties extends LazyLogging {
   val CorrelationIdKey = "correlationID"
   val CorrelationIdHeader = "X-Correlation-ID"
   val HealthControllerPath = "/health"
-  val ImageControllerPath = s"/image-api/v1/images"
+  val ImageApiBasePath = "/image-api/v1"
+  val ImageControllerPath = s"$ImageApiBasePath/images"
+  val RawControllerPath = s"$ImageApiBasePath/raw"
 
   val IsoMappingCacheAgeInMs = 1000 * 60 * 60 // 1 hour caching
   val LicenseMappingCacheAgeInMs = 1000 * 60 * 60 // 1 hour caching
