+++ src/test/scala/no/ndla/imageapi/TestEnvironment.scala
@@ -11,7 +11,7 @@ package no.ndla.imageapi
 
 import com.amazonaws.services.s3.AmazonS3Client
 import io.searchbox.client.JestClient
-import no.ndla.imageapi.controller.{HealthController, ImageController, InternController}
+import no.ndla.imageapi.controller.{HealthController, ImageController, InternController, RawController}
 import no.ndla.imageapi.integration._
 import no.ndla.imageapi.repository._
 import no.ndla.imageapi.service._
@@ -35,6 +35,7 @@ trait TestEnvironment
     with NdlaClient
     with InternController
     with ImageController
+    with RawController
     with TagsService
     with HealthController
     with ImageConverter
@@ -53,6 +54,7 @@ trait TestEnvironment
   val ndlaClient = mock[NdlaClient]
   val migrationApiClient = mock[MigrationApiClient]
   val imageController = mock[ImageController]
+  val rawController = mock[RawController]
   val internController = mock[InternController]
   val healthController = mock[HealthController]
   val converterService = mock[ConverterService]
